Create fetch query for license key by subscription
