+++ client/data/jetpack-licensing/types.ts
@@ -1,0 +1,13 @@
+export type UserLicenseApi = {
+	license_key: string;
+	product: string;
+	product_id: number;
+	subscription_id: number;
+};
+
+export type UserLicense = {
+	licenseKey: string;
+	product: string;
+	productId: number;
+	subscriptionId: number;
+};
