+++ client/data/jetpack-licensing/use-user-license-by-receipt-query.ts
@@ -1,28 +1,7 @@
 import { useQuery, UseQueryResult } from 'react-query';
 import wpcom from 'calypso/lib/wp';
-
-export type UserLicenseApi = {
-	license_key: string;
-	product: string;
-	product_id: number;
-	subscription_id: number;
-};
-
-export type UserLicense = {
-	licenseKey: string;
-	product: string;
-	productId: number;
-	subscriptionId: number;
-};
-
-function selectUserLicenseApiData( licenses: UserLicenseApi[] ): UserLicense[] {
-	return licenses.map( ( license ) => ( {
-		licenseKey: license.license_key,
-		product: license.product,
-		productId: license.product_id,
-		subscriptionId: license.subscription_id,
-	} ) );
-}
+import { mapManyLicenseApiToLicense } from './utils';
+import type { UserLicenseApi, UserLicense } from './types';
 
 const useUserLicenseByReceiptQuery = ( receiptId: number ): UseQueryResult< UserLicense[] > => {
 	const queryKey = [ 'user-license', receiptId ];
@@ -34,7 +13,9 @@ const useUserLicenseByReceiptQuery = ( receiptId: number ): UseQueryResult< User
 				apiNamespace: 'wpcom/v2',
 			} ),
 		{
-			select: selectUserLicenseApiData,
+			select: mapManyLicenseApiToLicense,
+			refetchIntervalInBackground: false,
+			refetchOnWindowFocus: false,
 		}
 	);
 };
