+++ client/data/jetpack-licensing/utils.ts
@@ -1,0 +1,12 @@
+import type { UserLicenseApi, UserLicense } from './types';
+
+export const mapSingleLicenseApiToLicense = ( license: UserLicenseApi ): UserLicense => ( {
+	licenseKey: license.license_key,
+	product: license.product,
+	productId: license.product_id,
+	subscriptionId: license.subscription_id,
+} );
+
+export const mapManyLicenseApiToLicense = ( licenses: UserLicenseApi[] ): UserLicense[] => {
+	return licenses.map( mapSingleLicenseApiToLicense );
+};
